rename realm_js_helpers.h to just realm_helpers.h
It isn't intended to be JS specific.
