+++ packages/bindgen/src/templates/node.ts
@@ -907,7 +907,7 @@ export function generate({ spec, file: makeFile }: TemplateContext): void {
 
   out(`
       #include <napi.h>
-      #include <realm_js_helpers.h>
+      #include <realm_helpers.h>
       #include <realm_js_node_helpers.h>
 
       namespace realm::js::node {
